Load only the slimmed down players.json initially
Everything else, including the transactions the players are involved in
will be loaded as needed. This should fix the JSON parsing errors in
Firefox, plus it will trade the long loading time for many short loads,
which should be a net win for the user since they aren't going to graph
all 60k players.