Call finish_func if there are no files to fetch
This was causing the George Bell trade tree to fail because one of the
players didn't have any other files to get, so it wasn't being
continued. This could cause problems if a large number of them happen
sequentially because now I'm developing a stack. I highly doubt that
will happen however and it should be easy to workaround if necessary.